src/main/java/dev/armancodeblock/springtdd/service/ProductServiceImpl.java
package dev.armancodeblock.springtdd.service;

import dev.armancodeblock.springtdd.entity.Product;
import dev.armancodeblock.springtdd.repository.ProductRepository;

import java.util.List;
import java.util.Optional;

public class ProductServiceImpl implements ProductService {

    private final ProductRepository productRepository;

    public ProductServiceImpl(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    @Override
    public Product save(Product product) {
        return productRepository.save(product);
    }

    @Override
    public Optional<Product> findById(Integer id) {
        return Optional.empty();
    }

    @Override
    public List<Product> findAll() {
        return List.of();
    }

    @Override
    public void deleteById(Integer id) {

    }

    @Override
    public Product updateById(Integer id,Product product) {
        return null;
    }
}

